hook up keywords and algorithm with controller

+++ grails-app/domain/seminario/domain/Status.groovy
@@ -2,10 +2,10 @@ package seminario.domain
 
 class Status {
 
-    Float accuracy
+    Integer accuracy
     String prediction
 
-    Status(Float accuracy, String prediction) {
+    Status(Integer accuracy, String prediction) {
         this.accuracy = accuracy
         this.prediction = prediction
     }
